Add obstruction helper module
